Split up pack tests, and rename fixture classes to avoid ambiguity.
